+++ R/ols-sandwich-var.R
@@ -11,11 +11,13 @@
 #' @param mod_fit A \code{\link[stats]{lm}} (OLS) object.
 #'
 #' @return A list containing the following elements: the type of estimator of
-#'   of the variance  (\code{var_type}); the summary statistics of \code{mod_fit}
-#'   based on this estimator of the variance (e.g., standard errors and p-values)
-#'   (\code{var_summary}); the assumptions under which the estimator of the
-#'   variance is consistent (\code{var_assumptions}); the covariance matrix for
-#'   the coefficients estimates (\code{cov_mat}).
+#'   of the variance  (\code{var_type}); An abbreviated string representing the
+#'   type of the estimator of the variance  (\code{var_type_abb}); the summary
+#'   statistics of \code{mod_fit} based on this estimator of the variance
+#'   (e.g., standard errors and p-values) (\code{var_summary}); the assumptions
+#'   under which the estimator of the variance is consistent
+#'   (\code{var_assumptions}); the covariance matrix for the coefficients
+#'   estimates (\code{cov_mat}).
 #'
 #' @keywords internal
 #'
@@ -39,9 +41,10 @@ comp_sand_var <- function(mod_fit) {
   )
   J_inv <- stats::summary.lm(mod_fit)$cov.unscaled
   X <- qr.X(mod_fit$qr)
-  V <- t(X) %*% Matrix::Diagonal(x = stats::residuals(mod_fit)^2) %*% X
+  meat <- t(X) %*% Matrix::Diagonal(x = stats::residuals(mod_fit)^2) %*% X
+  cov_mat <- as.matrix(J_inv %*% meat %*% J_inv)
 
-  std_error_sand <- sqrt(diag(as.matrix(J_inv %*% V %*% J_inv))) %>%
+  std_error_sand <- sqrt(diag(cov_mat)) %>%
     tibble::enframe(
       x = .,
       name = "term",
@@ -66,9 +69,10 @@ comp_sand_var <- function(mod_fit) {
     )
 
   out <- list(var_type = "sand",
+              var_type_abb = "sand",
               var_summary =  summary_sand,
-              var_assumptions = "The observations need to be i.i.d.",
-              cov_mat = V)
+              var_assumptions = "The observations must be i.i.d.",
+              cov_mat = cov_mat)
 
   return(out)
 }